+++ include/LabSound/extended/RealtimeAnalyser.h
@@ -59,7 +59,7 @@ private:
 
     // The audio thread writes the input audio here.
     AudioFloatArray m_inputBuffer;
-    unsigned m_writeIndex;
+    size_t m_writeIndex;
     
     uint32_t m_fftSize;
     std::unique_ptr<FFTFrame> m_analysisFrame;
